backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
click==8.2.1
exceptiongroup==1.3.0
fastapi==0.104.1
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
podman-compose==1.5.0
psycopg2-binary==2.9.10
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.2
scapy==2.6.1
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.27.0
typing_extensions==4.15.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
redis>=4.5.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
requests>=2.28.0
python-decouple>=3.6
